feat(app): New flag for deleting Neo4j data added. Each Neo4j node is given a 'name' property if it doesn't exist
A flag to delete data from Neo4j is added. All indices, constraints, nodes and relationships are deleted (that represent the databases defined in .env).
To enhance visualization, a 'name' property is created for those nodes that do not have one. In the browser, the nodes display the 'name' property